# feat(core): Replace `anyio.to_thread.run_sync` with native versions