--- a/docs/reference/concurrency.rst
+++ b/docs/reference/concurrency.rst
@@ -1,0 +1,5 @@
+cli
+===
+
+.. automodule:: litestar.concurrency
+    :members: