- Replace `anyio.to_thread.run_sync` with native versions of supported libraries, i.e. [`asyncio.loop.run_in_executor`](https://docs.python.org/3/library/asyncio-eventloop.html#asyncio.loop.run_in_executor) and [`trio.to_thread.run_sync`](https://trio.readthedocs.io/en/stable/reference-core.html#trio.to_thread.run_sync)
- Add `set_asyncio_executor` and `get_asyncio_executor` utility functions to configure the executor used when running synchronous functions
- Add `set_trio_capacity_limiter` and `get_trio_capacity_limiter` utility functions to configure the `trio.CapacityLimiter` used when running synchronous functions

<hr>

Reasons for this change are twofold:

1. Performance improvements. Anyio has some [known performance overhead](https://github.com/agronholm/anyio/discussions/543), and its threading solution proved to be slower than both asyncio's and trio's built-in functionalities
2. Configurability of the thread pools used under the hood. Prior to this change the only way a user could influence these was to [set a global default in anyio](https://anyio.readthedocs.io/en/stable/threads.html#adjusting-the-default-maximum-worker-thread-count)

<hr>

Using this very simple benchmark we can see the stark difference in performance between anyio and native asyncio:

```python
import asyncio
import time
from concurrent.futures import ThreadPoolExecutor
from timeit import timeit

from anyio.to_thread import run_sync

from litestar.concurrency import sync_to_thread, set_asyncio_executor

set_asyncio_executor(ThreadPoolExecutor(40))


def func():
    time.sleep(0)


async def run_anyio():
    await run_sync(func)


async def run_native():
    await sync_to_thread(func)


async def run_anyio_batch():
    await asyncio.gather(*[run_sync(func) for _ in range(100)])


async def run_native_batch():
    await asyncio.gather(*[sync_to_thread(func) for _ in range(100)])


def run(fn):
    asyncio.new_event_loop().run_until_complete(fn())


print(timeit("run(run_anyio)", globals=globals(), number=1000))
print(timeit("run(run_native)", globals=globals(), number=1000))

print(timeit("run(run_anyio_batch)", globals=globals(), number=100))
print(timeit("run(run_native_batch)", globals=globals(), number=100))
```

```shell
0.6600718750014494
0.19882075799978338
2.053338443998655
0.7532648650012561
``` 

And testing this in a simple application, using `wrk`:

```python
from litestar import get, Litestar

@get(sync_to_thread=True)
def handler() -> str:
    return "Hello, world!"

app = Litestar(route_handlers=[handler])
```

*Without this PR*

```text
Thread Stats    Avg      Stdev     Max     +/- Stdev
Latency       23.83ms   14.72ms  266.88ms   98.07%
Req/Sec        2.22k    217.67    2.39k     95.00%

22114 requests in 5.01s, 3.10MB read
Requests/sec:   4415.57
Transfer/sec:   633.88KB
```

*With the changes applied*

```text
Thread Stats    Avg      Stdev     Max     +/- Stdev
Latency       17.14ms   13.38ms  254.68ms   96.68%
Req/Sec        3.14k     397.07   3.66k     83.00%

31234 requests in 5.01s, 4.38MB read
Requests/sec:   6236.19
Transfer/sec:   0.87MB
```