@provinzkraut 2023-12-31
Two more things to consider regarding anyio would be file operations (which we're not using that frequently) and `TaskGroup`s, which we *do* use in quite a few places. A quick benchmarks shows the native `asyncio.TaskGroup` to be about twice as performant as the one provided by anyio. This is not an entirely fair comparison though as the anyio's task group does a lot more. It would be possible to provide a backport for the native `TaskGroup` for Python < 3.11 (it's not a lot of code actually) and only fall back on anyio if we're not running on asyncio. 

This poses a general question: Do we want to use anyio for the features it provides, because we want to make use of its structured concurrency approach, or just as a shim to provide asyncio/trio compatibility?

Our trio compatibility is spotty in general and not particularly well tested and dropping it has come up a few times now. On that front it should also be noted that currently the only ASGI server I'm aware of that supports trio is hypercorn, and major libraries like redis or SQLAlchemy don't support it, so the potential audience for it is limited to begin with.

Personally, I'm not too keen to start introducing our own compatibility layer between asyncio/trio, but as it stands, we're not really using anyio for anything more than that aside from a few utilities we could easily implement ourselves. I think if we want to keep anyio, we should actually embrace it and the structured concurrency approach, and if we're not going that route, it makes sense to fall back on pure asyncio to gain some performance, simplify our implementation and remove external dependencies.

@litestar-org/maintainers