@FHU-yezi 2024-01-01
I think we can remove `anyio` as welll as our `trio` support in next major version, because the community have poor support in developing API with trio.