@provinzkraut 2024-01-02
> > This poses a general question: Do we want to use anyio for the features it provides, because we want to make use of its structured concurrency approach, or just as a shim to provide asyncio/trio compatibility?
> 
> speaking for myself, I never use trio so not really interested in the asyncio/trio compatibility b but structured concurrency approach is defintely a plus. would quattro help ? (https://github.com/Tinche/quattro)

Interesting, hadn't seen that library yet! Looking at it, it seems to be just a few hundred lines of code, half of which is the backport of `asyncio.TaskGroup`. Maybe it's worth considering to just implement what we need ourselves!